+++ lib/packages/wikibase/data-model/src/Entity/EntityDocument.php
@@ -70,6 +70,9 @@ interface EntityDocument extends Comparable {
 
 	/**
 	 * Returns a deep clone of the entity. The clone must be equal in all details, including the id.
+	 * No change done to the clone is allowed to interfere with the original object. Only properties
+	 * containing immutable objects are allowed to (and should) reference the original object.
+	 *
 	 * Since EntityDocuments are not immutable (at least the id can be set) the method is not
 	 * allowed to return $this.
 	 *