+++ lib/packages/wikibase/data-model/src/Entity/Property.php
@@ -259,7 +259,17 @@ class Property extends Entity implements StatementListHolder {
 	 * @return self
 	 */
 	public function copy() {
-		return unserialize( serialize( $this ) );
+		return clone $this;
+	}
+
+	/**
+	 * @see http://php.net/manual/en/language.oop5.cloning.php
+	 *
+	 * @since 5.1
+	 */
+	public function __clone() {
+		$this->fingerprint = clone $this->fingerprint;
+		$this->statements = clone $this->statements;
 	}
 
 }